Upgrade espresso to 2.0. Switch to android-unit-test plugin. Update README
